- added algorithm to calculate servo angle from joint angle;

+++ pc_cpp/src/Point3d.cpp
@@ -2,23 +2,23 @@
 
 class Point3d {
 public:
-    double x, y, z;
+    float x, y, z;
     Point3d() {
         x = y = z = 0.0f;
     }
-    Point3d(double x, double y, double z) {
+    Point3d(float x, float y, float z) {
         this->x = x;
         this->y = y;
         this->z = z;
     }
-    static double distance_between(Point3d* p1, Point3d* p2) {
-        double diffX = p1->x - p2->x;
-        double diffY = p1->y - p2->y;
-        double diffZ = p1->z - p2->z;
+    static float distance_between(Point3d* p1, Point3d* p2) {
+        float diffX = p1->x - p2->x;
+        float diffY = p1->y - p2->y;
+        float diffZ = p1->z - p2->z;
         return sqrt(diffX*diffX + diffY*diffY + diffZ*diffZ);
     }
 
-    double distance_to(Point3d* other) {
+    float distance_to(Point3d* other) {
         return distance_between(this, other);
     }
 };
